backend/requirements.txt
# SRIMCA AI Backend Dependencies

# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB Driver
pymongo==4.6.1
dnspython==2.4.2

# Environment Variables
python-dotenv==1.0.0

# JSON Handling
requests==2.31.0

# Password Hashing
bcrypt==4.1.2

# JWT Authentication
PyJWT==2.8.0

# Date/Time Handling
python-dateutil==2.8.2

# Firebase Admin SDK
firebase-admin==6.3.0

# Gunicorn for production deployment
gunicorn==21.2.0

# Python-dotenv for loading .env files
python-dotenv==1.0.0
